precompute: fix bf16->numpy conversion in the tiled VAE path (PRECOMP-1 bug)
The tiled VAE encode (>512px) returned np.array(lat) on a bfloat16 mlx array,
which numpy rejects ("Item size 2 for PEP 3118 buffer format string B") — every
record at 768/1024px was skipped, so the high-res tiling path (landed 2026-05-22)
had never actually produced a latent. The <=512 batch path and the qwen3 path
already cast (np.array(x.astype(mx.float32))); apply the same cast at both tiled
return sites. Found by the TRAIN-7 768px probe precompute — the same May-window
untested-path pattern as QWEN-1/MLX-1.

+++ train/scripts/precompute_all.py
@@ -246,7 +246,8 @@ def _encode_vae_tiled(vae, img_chw: np.ndarray,
         x = mx.array(img_chw[np.newaxis])  # [1, 3, H, W]
         lat = vae.encode(x)
         mx.eval(lat)
-        return np.array(lat)[0]  # [32, H//8, W//8]
+        # bf16 -> f32 before numpy: bfloat16 has no numpy dtype (PEP 3118 'B' error)
+        return np.array(lat.astype(mx.float32))[0]  # [32, H//8, W//8]
 
     result  = np.zeros((lC, lH, lW), dtype=np.float32)
     weights = np.zeros((lH, lW), dtype=np.float32)
@@ -268,7 +269,7 @@ def _encode_vae_tiled(vae, img_chw: np.ndarray,
             x_t = mx.array(tile[np.newaxis])  # [1, 3, tile_size, tile_size]
             lat = vae.encode(x_t)
             mx.eval(lat)
-            lat_np = np.array(lat)[0]  # [32, lT, lT]
+            lat_np = np.array(lat.astype(mx.float32))[0]  # [32, lT, lT]
 
             # Valid latent region (clip padded edge)
             ly0, ly1 = y0 // 8, y1 // 8
